Refresh private FeatherPGS API
Port Zach's heterogeneous-world FeatherPGS fixes, replace the tiled kernel factory class with cached module-level helpers, consolidate the FeatherPGS tests, and expand private solver docstrings for review.
